(halium) libaudioclient: Avoid stalling audio recording on SCHED_* checks
libaudioclient (responsible for polling buffers from micshm) waits up to half
a second on checking whether real-time scheduling priority has been set or not.

This has the unfortunate side-effect of stalling the audio record pipeline
for as long as the check is done, over and over again, causing A/V desync.

Remove the check altogether as there is no reason to stall audio recording
while MPEG4Writer is doing its job fine. If the priority is set, it's set.
If not, tough.